move connectors to module
