+++ src/connectors/mod.rs
@@ -1,0 +1,7 @@
+mod connector;
+mod insecure_connector;
+mod secure_connector;
+
+pub use connector::Connector;
+pub use insecure_connector::InsecureConnector;
+pub use secure_connector::SecureConnector;
